+++ CHANGELOG.md
@@ -11,4 +11,5 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Distribute operator.
 - Empty parent operator.
 - Mesh name fixer/matcher operator.
+- LOD creation operator.
 - Bulk FBX export operator.